+++ src/utils/copy/copy_dense.cpp
@@ -28,5 +28,21 @@ template <typename T> void Dense<T>::operator=(const Dense<T> &mat) {
 template void Dense<double>::operator=(const Dense<double> &mat);
 template void Dense<float>::operator=(const Dense<float> &mat);
 
+template <typename T>
+void Dense<T>::set_ptr(const size_t M, const size_t N, const std::vector<T> &value) {
+  Logger &logger = Logger::get_instance();
+  logger.util_in(monolish_func);
+  val = value;
+
+  rowN = M;
+  colN = N;
+  nnz = M*N;
+  logger.util_out();
+}
+template void Dense<double>::set_ptr(const size_t M, const size_t N,
+                                   const std::vector<double> &value);
+template void Dense<float>::set_ptr(const size_t M, const size_t N,
+                                  const std::vector<float> &value);
+
 } // namespace matrix
 } // namespace monolish