+++ src/utils/crs_constructor.cpp
@@ -66,7 +66,7 @@ CRS<T>::CRS(const size_t M, const size_t N, const size_t NNZ, const int *rowptr,
 
 #pragma omp parallel for
   for (size_t i = 0; i < nnz; i++) {
-    col_index[i] -= origin;
+    col_ind[i] -= origin;
   }
 
   compute_hash();
@@ -121,6 +121,21 @@ CRS<T>::CRS(const size_t M, const size_t N, const std::vector<int> &rowptr,
   std::copy(rowptr.data(), rowptr.data() + (M + 1), row_ptr.begin());
   std::copy(colind.data(), colind.data() + nnz, col_ind.begin());
   std::copy(value.data(), value.data() + nnz, val.begin());
+
+  if(value.get_device_mem_stat() == true){
+#if MONOLISH_USE_GPU
+  send();
+  const T* data = value.data();
+#pragma omp target teams distribute parallel for
+    for (size_t i = 0; i < get_nnz(); i++) {
+      val[i] = data[i];
+    }
+#else
+    throw std::runtime_error(
+        "error USE_GPU is false, but get_device_mem_stat() == true");
+#endif
+  }
+
   compute_hash();
   logger.util_out();
 }